Feature/fix subcription v2 (#77)
* Fix Caculator LLM

* Add more status in subcription

* History Payment Dept

* pull default

* Fix Subcription Quota

* Fix test migration

* Pull default

+++ src/main/java/com/sep/vox/domain/dto/PlanQuotaDto.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public record PlanQuotaDto(
     UUID id,
     String quotaType,
-    Integer includedQuantity,
+    BigDecimal includedQuantity,
     BigDecimal tokenUnitPrice
 ) {
 }
