+++ src/main/java/com/sep/vox/domain/dto/SubscriptionQuotaUserAllocationDto.java
@@ -1,12 +1,13 @@
 package com.sep.vox.domain.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record SubscriptionQuotaUserAllocationDto(
     UUID userId,
     String fullName,
     String quotaType,
-    Integer allocatedQuantity,
-    Integer usedQuantity
+    BigDecimal allocatedQuantity,
+    BigDecimal usedQuantity
 ) {
 }
