+++ src/main/java/com/sep/vox/domain/dto/TokenUsageTimeseriesDto.java
@@ -1,10 +1,11 @@
 package com.sep.vox.domain.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record TokenUsageTimeseriesDto(
     String granularity,
-    Integer totalUsed,
+    BigDecimal totalUsed,
     List<TokenUsageTimeseriesPointDto> points,
     List<SubscriptionQuotaDto> currentPeriod
 ) {
